refactor: 입/출력 로직을 StringHanlder로부터 각각 분리
