+++ src/main/java/Main.java
@@ -1,12 +1,14 @@
-package calculator;
+import calculator.Formula;
+import view.InputView;
+import view.OutputView;
 
 public class Main {
     public static void main(String[] args) {
         while (true) {
             try {
-                String[] inputs = StringHandler.getInputAndSplit();
+                String[] inputs = InputView.getInputAndSplit();
                 Formula formula = new Formula(inputs);
-                StringHandler.printResult(formula.calculateFormula());
+                OutputView.printResult(formula.calculateFormula());
                 return;
             } catch (IllegalArgumentException e) {
                 e.printStackTrace();
