+++ src/main/java/view/OutputView.java
@@ -1,0 +1,7 @@
+package view;
+
+public class OutputView {
+    public static void printResult(double result) {
+        System.out.println("result: " + result);
+    }
+}
